pyproject.toml
[tool.poetry]
name = "pollka"
version = "0.1.0"
description = ""
authors = ["EliseyGodX"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.1"
djangorestframework = "^3.16.0"
djangorestframework-simplejwt = "^5.5.0"
django-debug-toolbar = "^5.2.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
flake8-bugbear = "^24.12.12"
flake8-docstrings = "^1.7.0"
flake8-isort = "^6.1.1"
flake8-pyproject = "^1.2.3"
mccabe = "^0.7.0"
pycodestyle = "^2.12.1"
pyflakes = "^3.2.0"
unify = "^0.5"
flake8-in-file-ignores = "^0.3.0"
wemake-python-styleguide = "1.0"

[tool.flake8]
max-line-length = 88
required-plugins = [
    "flake8-bugbear",
    "flake8-docstrings",
    "flake8-isort",
    "flake8-pyproject",
    "mccabe",
    "pycodestyle",
    "pyflakes",
    "flake8-in-file-ignores",
]

ignore = [
    "ANN002",
    "B907",
    "D100",
    "D101",
    "D102",
    "D103",
    "D105",
    "D104",
    "D106",
    "D107",
    "E701",
    "E704",
    "I001",
    "I005",
    "W503",
    "WPS111",
    "WPS201",
    "WPS202",
    "WPS210",
    "WPS214",
    "WPS217",
    "WPS226",
    "WPS229",
    "WPS235",
    "WPS504",
    "WPS602",
    "WPS604",
    "WPS420",
    "WPS430",
    "WPS231",
    "WPS232",
    "WPS238",
    "WPS211",
    "WPS336",
    "WPS432",
    "WPS110"
]

select = ["B", "C", "D", "E", "F", "I", "T4", "W"]
extend-select = ["B9"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
